Moved typechecked specs into separate crate
