+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "secret-integers",
     "abstract-integers",
     "spec-examples",
+    "hacspec-examples",
     "hacspec-attributes",
     "hacspec-sig"
 ]
@@ -17,6 +18,7 @@ default-members = [
     "secret-integers",
     "abstract-integers",
     "spec-examples",
+    "hacspec-examples",
     "hacspec-attributes",
     "hacspec-sig"
 ]
