+++ hacspec-examples/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "hacspec-examples"
+name = "hacspec-examples-typechecked"
 version = "0.1.0"
-authors = ["Franziskus Kiefer <franziskuskiefer@gmail.com>"]
+authors = ["Denis Merigoux <denis.merigoux@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -11,7 +11,6 @@ hacspec-lib = { path = "../hacspec-lib" }
 hacspec-derive = { path = "../hacspec-derive" }
 secret_integers = { path = "../secret-integers" }
 abstract_integers = { path = "../abstract-integers" }
-hacspec-examples-typechecked = { path = "../hacspec-examples" }
 
 [dev-dependencies]
 # FIXME: Why?
