+++ hacspec-examples/src/chacha20_poly1305/mod.rs
@@ -1,0 +1,1 @@
+pub mod chacha20;
