+++ hacspec-examples/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod chacha20_poly1305;
