+++ spec-examples/src/chacha20_poly1305/mod.rs
@@ -1,4 +1,3 @@
-pub mod chacha20;
 mod chacha20poly1305;
 pub mod poly1305;
